internal: move dateFormatForLiquibase to base and rename to nextTimestamp
