+++ generators/base-core/generator.ts
@@ -40,7 +40,6 @@ import {
   CRLF,
   LF,
   createJHipster7Context,
-  formatDateForChangelog,
   hasCrlr,
   joinCallbacks,
   normalizeLineEndings,
@@ -75,7 +74,6 @@ import { extractArgumentsFromConfigs } from '../../lib/command/index.js';
 import type BaseApplicationGenerator from '../base-application/generator.js';
 import type { Config } from '../base-core/types.js';
 import type { GenericTaskGroup } from '../../lib/types/base/tasks.js';
-import { CONTEXT_DATA_REPRODUCIBLE_TIMESTAMP } from '../base-application/support/constants.js';
 import { convertWriteFileSectionsToBlocks } from './internal/index.js';
 
 const {
@@ -583,54 +581,6 @@ You can ignore this error by passing '--skip-checks' to jhipster command.`);
           storage,
         };
       });
-  }
-
-  /**
-   * Generate a date to be used by Liquibase changelogs.
-   *
-   * @param {Boolean} [reproducible=true] - Set true if the changelog date can be reproducible.
-   *                                 Set false to create a changelog date incrementing the last one.
-   * @return {String} Changelog date.
-   */
-  dateFormatForLiquibase(reproducible?: boolean): string {
-    reproducible = reproducible ?? Boolean(this.options.reproducible);
-    // Use started counter or use stored creationTimestamp if creationTimestamp option is passed
-    const creationTimestamp = this.options.creationTimestamp ? this.config.get('creationTimestamp') : undefined;
-    let now = new Date();
-    // Miliseconds is ignored for changelogDate.
-    now.setMilliseconds(0);
-    // Run reproducible timestamp when regenerating the project with reproducible option or an specific timestamp.
-    if (reproducible || creationTimestamp) {
-      now = this.getContextData(CONTEXT_DATA_REPRODUCIBLE_TIMESTAMP, {
-        factory: () => {
-          const newCreationTimestamp: string = (creationTimestamp as string) ?? this.config.get('creationTimestamp');
-          const newDate = newCreationTimestamp ? new Date(newCreationTimestamp) : now;
-          newDate.setMilliseconds(0);
-          return newDate;
-        },
-      });
-      now.setMinutes(now.getMinutes() + 1);
-      this.getContextData(CONTEXT_DATA_REPRODUCIBLE_TIMESTAMP, { override: now });
-
-      // Reproducible build can create future timestamp, save it.
-      const lastLiquibaseTimestamp = this.jhipsterConfig.lastLiquibaseTimestamp;
-      if (!lastLiquibaseTimestamp || now.getTime() > lastLiquibaseTimestamp) {
-        this.config.set('lastLiquibaseTimestamp', now.getTime());
-      }
-    } else {
-      // Get and store lastLiquibaseTimestamp, a future timestamp can be used
-      const lastLiquibaseTimestamp = this.jhipsterConfig.lastLiquibaseTimestamp;
-      if (lastLiquibaseTimestamp) {
-        const lastTimestampDate = new Date(lastLiquibaseTimestamp);
-        if (lastTimestampDate >= now) {
-          now = lastTimestampDate;
-          now.setSeconds(now.getSeconds() + 1);
-          now.setMilliseconds(0);
-        }
-      }
-      this.jhipsterConfig.lastLiquibaseTimestamp = now.getTime();
-    }
-    return formatDateForChangelog(now);
   }
 
   /**
