+++ generators/base/generator.ts
@@ -32,11 +32,15 @@ import type { TaskTypes as BaseTaskTypes, GenericTaskGroup } from '../../lib/typ
 import type { Config } from '../base-core/types.js';
 import { CONTEXT_DATA_EXISTING_PROJECT } from '../base-application/support/constants.js';
 import { GENERATOR_JHIPSTER } from '../generator-constants.js';
-import { packageNameToNamespace } from './support/index.js';
+import { formatDateForChangelog, packageNameToNamespace } from './support/index.js';
 import { loadBlueprintsFromConfiguration, mergeBlueprints, normalizeBlueprintName, parseBluePrints } from './internal/index.js';
 import { PRIORITY_NAMES } from './priorities.js';
 import type { JHipsterGeneratorFeatures, JHipsterGeneratorOptions } from './api.js';
-import { CONTEXT_DATA_BLUEPRINT_CONFIGURED, LOCAL_BLUEPRINT_PACKAGE_NAMESPACE } from './support/constants.js';
+import {
+  CONTEXT_DATA_BLUEPRINT_CONFIGURED,
+  CONTEXT_DATA_REPRODUCIBLE_TIMESTAMP,
+  LOCAL_BLUEPRINT_PACKAGE_NAMESPACE,
+} from './support/constants.js';
 import type { CleanupArgumentType, Control } from './types.js';
 
 const { WRITING } = PRIORITY_NAMES;
@@ -233,6 +237,50 @@ export default class JHipsterBaseBlueprintGenerator<
         };
       },
     });
+  }
+
+  /**
+   * Generate a timestrap to be used by Liquibase changelogs.
+   */
+  nextTimestamp(): string {
+    const reproducible = Boolean(this.options.reproducible);
+    // Use started counter or use stored creationTimestamp if creationTimestamp option is passed
+    const creationTimestamp = this.options.creationTimestamp ? this.config.get('creationTimestamp') : undefined;
+    let now = new Date();
+    // Miliseconds is ignored for changelogDate.
+    now.setMilliseconds(0);
+    // Run reproducible timestamp when regenerating the project with reproducible option or an specific timestamp.
+    if (reproducible || creationTimestamp) {
+      now = this.getContextData(CONTEXT_DATA_REPRODUCIBLE_TIMESTAMP, {
+        factory: () => {
+          const newCreationTimestamp: string = (creationTimestamp as string) ?? this.config.get('creationTimestamp');
+          const newDate = newCreationTimestamp ? new Date(newCreationTimestamp) : now;
+          newDate.setMilliseconds(0);
+          return newDate;
+        },
+      });
+      now.setMinutes(now.getMinutes() + 1);
+      this.getContextData(CONTEXT_DATA_REPRODUCIBLE_TIMESTAMP, { override: now });
+
+      // Reproducible build can create future timestamp, save it.
+      const lastLiquibaseTimestamp = this.jhipsterConfig.lastLiquibaseTimestamp;
+      if (!lastLiquibaseTimestamp || now.getTime() > lastLiquibaseTimestamp) {
+        this.config.set('lastLiquibaseTimestamp', now.getTime());
+      }
+    } else {
+      // Get and store lastLiquibaseTimestamp, a future timestamp can be used
+      const lastLiquibaseTimestamp = this.jhipsterConfig.lastLiquibaseTimestamp;
+      if (lastLiquibaseTimestamp) {
+        const lastTimestampDate = new Date(lastLiquibaseTimestamp);
+        if (lastTimestampDate >= now) {
+          now = lastTimestampDate;
+          now.setSeconds(now.getSeconds() + 1);
+          now.setMilliseconds(0);
+        }
+      }
+      this.jhipsterConfig.lastLiquibaseTimestamp = now.getTime();
+    }
+    return formatDateForChangelog(now);
   }
 
   /**
