+++ generators/base/support/constants.ts
@@ -1,2 +1,3 @@
 export const LOCAL_BLUEPRINT_PACKAGE_NAMESPACE = '@jhipster/generator-jhipster-local';
 export const CONTEXT_DATA_BLUEPRINT_CONFIGURED = 'jhipster:blueprintConfigured';
+export const CONTEXT_DATA_REPRODUCIBLE_TIMESTAMP = 'jhipster:reproducibleTimestamp';
